requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
Jinja2==3.1.4
libcst==1.4.0
MarkupSafe==2.1.5
pyspellchecker==0.8.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
